Use minor-mode and text properties instead of link
This is a different approach to jumping to source notes in the
backlinks buffer. Rather than constructing an org-mode link to the
source, this defines a new minor-mode for the backlinks buffer and
adds text properties to the backlink content snippets containing the
originating file and the snippet position in that file. The minor mode
has a keybinding (mouse-1/RET) that navigates to the source file and
goes to the snippet position.