Increase performance of post-command-hook
file-truename can be an expensive function on a slow filesystem like NFS. I
Removed a lot of the unneeded code and refactored to improve performance. In my
testing it took the execution time from 13ms per call to 2µs; over a 1000x
speedup. This is important since post-command-hook is called with every
character you type.