Add org-ref link support
This commit adds support for org-ref links of the form
"cite:bibtex-key". The code assumes that the notes for a paper are
stored in
/org/roam/directory/<bibtex-key>.org

The function `org-element-parse-buffer' had to be modified quite
extensively because an org-ref link can link to multiple papers at
once. In these multi-links, the bibtex keys are separated by commas,
e.g., "cite:zhu-2017-unpair-image,isola-2016-image-tokl".

I have added the function `org-roam--link-get-content' that obtains
the content of a link. This function is used when handling both file
links and org-ref style links, thus minimizing code repetition.

Previously, the property `:point' of an org-roam item would be
determined by the `:content-begin' or `:begin' property of the
`org-element-at-point'. Now, the `:point' property of an org-roam item
is populated by the `:begin' property of the link. I do not know and
have not tested if this leads to observably different behaviour.

I have added a dependency on dash.el because I wanted to use
`-flatten-n'.