Change develop to master in installation
As a newcomer, I spent hours trying to figure out why links didn't format properly, until I discovered that it was because I was on develop branch. Switching to master solved all my pains. I hope others don't have to figure it out in the future, maybe just put a disclaimer somewhere in the page :smiley: