Remove over-engineering
Actually, we don’t need to do any of that.  If ‘org-id-track-globally’ is
t (default), the links will be stored in ‘org-id-locations-file’ as expected
based on the hash-table ‘org-id-locations’.  If it’s nil, the hash-table is
not created nor updated, which means that the file is not created.  All good,
no need to nuke the behaviour of ‘org-id-add-location’.

This solution has the side-effects of storing links to Org-roam files in
‘org-id-locations-file’ though, and although we might be able to update it on
renames, it’ll be flimsy.  This only matters for external files linking to
headings inside the ‘org-roam-directory’, though, so it’s expected that we’re
not handling those.  Inside ‘org-roam-directory’, we’ll be using our own sql
database to jump around anyway.