Add prompt when reverting to ‘org-id-goto’
This is because ‘org-id-goto’ will parse the ‘org-id-files’ if the ID not
stored stored in its hash-table, which can takes a while if you have large
agenda-files.

I’m not sure whether there’s a better way to do this if we want to maintain
the default behaviour of IDs in Org-mode.  If we consider that a link inside
an Org-roam file pointing to an ID that does not exist in the
Org-roam-directory should be repaired, it prevents linking to ID outside of
‘org-roam-directory’.  This might make it harder to repair id-links without
destination, but since they are UUIDs, they should be even sturdier than
filenames.

The problem is more likely to be a proliferation of unreferenced :ID:
properties inside ‘org-roam-directory’.  We could clean them up, but that
might break links outside of ‘org-roam-directory’ pointing to a roam-headline.
There isn’t any reliable way to check that without parsing the
‘org-id-files’ (with included in ‘org-agenda-files’ via ‘org-id-extra-files’).
This could be hidden behind an EXHAUSTIVE argument for the checker.