Split build into two-step
This is needed because id-links cannot be inserted if the headline hasn’t
already been inserted in the db.  The one-step process we had considered files
linearily, which meant that some files with headlines were processed before
files with id-links pointing to them.

Now, the two-step expansion does a first pass to gather the files and the
headlines, and then a second pass to process the rest.  This adds a little bit
of computation, since we have to go through the list twice, but we’re
performing different functions on them, so it’s not entirely redundant.