Add org-roam--completion-exit-fn
Currently Org-roam's completion-at-point implements :exit-fn as
a lambda function. For some reason, it is not triggered at post-
completion; however, it has been observed that a function is
triggered as expected. This commit implements a function, and
replaces the assignment of :exit-function with the newly
created function. This is intended to get the cursor (point)
out of the [[Title]] bracket for more intuitive writing experience.