(feat): enable capt for [[|]]; add exit-fn
The intention of this commit / PR is to suggest an alternative to the current
capf in org-roam-link.el. Currently, you need to type [[roam:]] for capf to
work for the link completion-at-point. This alternative enables capf to be
triggered withoout roam: for [[|]]. roam: is added automatically with
exit-fn (implemented as a lambda function as originally intended).