(doc): Add footnote for Getting Started
Follow on to #1345.

The discussion in the Slack thread further clarified what was exactly
confusing for a beginner reader of this Getting Started. When a user is using
the built-in completion (no ivy, no ido), then `org-roam-find-file` does not
immediately show any files -- instead, the user has to press TAB. This is not
explicitly mentioned.

https://orgroam.slack.com/archives/CV160S8EL/p1608121018061700?thread_ts=1608039505.055000&cid=CV160S8EL