Inhibit org-capture from setting :exact-position by default
`org-capture-set-target-location` will always try to set
`:exact-position` for `function` based capture targets, which is what
`org-roam-capture--get-point` is.

The problem with this is that `:exact-position` doesn't play well with
various capture `type`s and it breaks `entry` based capture templates
from proper insertion. The major reason for this is a conflict between
`:exact-position` and `:target-entry-p` in org-capture, which
conditionally compete with each other[0].

This won't prohibit org-roam's based capture targets from setting
`:exact-position` for targets that would actually need it[1], but would
inhibit org-capture from setting the `:exact-position` by default.

[0]: See `org-capture-place-{entry,item,table-line,plain-text}` and
`org-capture-set-target-location` for how this is implemented.

[1]: This would be accomplished similarly to how
`org-capture-set-target-location` does so.