(refactor): centralize eval-when-compile and get rid of org-roam-macs.el
With the new load chain, just like in case of the dependencies, we can
simply keep all the `eval-when-compile` business where org-roam.el is.

org-roam-macs.el is redundant when we already have org-roam-utils.el and
in practive there's no observable benefits of splitting utility macros
from the rest of the facility. The macros (like before) still will be
expanded during byte compilation or eagerly during loading.

Side notes:
- No need to (require 'org-macs), since it will be already required by
  `org` and its libraries.
- No need to wrap (require 'org-roam-utils) with something like
  `eval-and-compile`, since top level `require` forms are already
  implicitly wrapped as such.