Stop point moving when visiting node if point already at node
When a buffer is buried and then brought up again, it is unexpected for
point to move. Currently, org-roam-node-find-noselect always sets point
to the pos column from the database, which corresponds to the very
beginning of the node in its file.

This patch modifies org-roam-node-find-noselect to check whether point
for the buffer associated with the node is already at the target
node. Only if it is not, is point moved to the target node.