(feat)capture: prefix arg forces capture in org-roam-node-insert
Add an argument FORCE-CAPTURE to `org-roam-node-insert', which is a
prefix arg when called interactively. If FORCE-CAPTURE is non-nil,
then capture to the selected node before inserting the link regardless
of whether it already exists (previous behaviour was to only capture
to new nodes, this remains the default without FORCE-CAPTURE).

Add a line to docstring explaining default behaviour.

Update docstring for new arg.